LayoutTests/platform/gtk/accessibility/title-and-alt-expected.txt


Functional label:
This tests the exposure of alt and title information.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS image.title is 'Image alt attr'
PASS image.description is 'Image title attr'
PASS link.title is ''
PASS link.description is 'Link title attr'
PASS entry.title is ''
PASS entry.description is 'Entry title attr'
PASS button.title is 'Submit'
PASS button.description is 'Submit button title attr'
PASS successfullyParsed is true

TEST COMPLETE
